Update D8 R8 to 8.12.7-dev
This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/8.12.7-dev

BYPASS_INCLUSIVE_LANGUAGE_REASON=Mapping file include names from third party libraries

Change-Id: I2df9a78a930d7e7d0abf4821bd384f5266ffb995